% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{cal_score_init}
\alias{cal_score_init}
\title{Calculate score for each feature in each cell}
\usage{
cal_score_init(
  expr,
  tf = c("logtf", "tf"),
  idf = "prob",
  iae = "prob",
  par.idf = NULL,
  par.iae = NULL,
  return.intermediate = FALSE
)
}
\arguments{
\item{expr}{a count matrix, features in row and cells in column}

\item{tf}{a character, specify the TF method to use, can be "tf" or "logtf"}

\item{idf}{a character, specify the IDF method to use. Available methods can
be accessed using \code{\link[=idf_iae_methods]{idf_iae_methods()}}}

\item{iae}{a character, specify the IAE method to use. Available methods can
be accessed using \code{\link[=idf_iae_methods]{idf_iae_methods()}}}

\item{par.idf}{other parameters for specified IDF methods}

\item{par.iae}{other parameters for specified IAE methods}

\item{return.intermediate}{logical, if TRUE the returned list also contains
the intermediate \code{tf}, \code{idf} and \code{iae} objects. Default \code{FALSE} keeps
only the combined \code{score} to avoid the memory overhead of three extra
feature-by-cell matrices on large inputs.}
}
\value{
a list always containing \code{score}; when \code{return.intermediate = TRUE}
the list additionally contains \code{tf}, \code{idf} and \code{iae}.
}
\description{
Calculate score for each feature in each cell
}
\examples{
data <- matrix(rpois(100, 2), 10, dimnames = list(1:10))
label <- sample(c("A", "B"), 10, replace = TRUE)
smartid:::cal_score_init(data,
  par.idf = list(label = label),
  par.iae = list(label = label)
)
}
