% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{gs_score_init}
\alias{gs_score_init}
\title{Calculate scores of each cell on given features}
\usage{
gs_score_init(score, features = NULL)
}
\arguments{
\item{score}{matrix, features in row and samples in column}

\item{features}{vector, feature names to compute score}
}
\value{
a vector of score
}
\description{
Calculate scores of each cell on given features
}
\examples{
data <- matrix(rnorm(100), 10, dimnames = list(1:10))
gs_score_init(data, 1:5)
}
