% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{idf_iae_methods}
\alias{idf_iae_methods}
\title{Get names of available IDF and IAE methods}
\usage{
idf_iae_methods()
}
\value{
names of methods implemented
}
\description{
Returns a named vector of IDF/IAE methods
}
\examples{
idf_iae_methods()
}
