% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{score_barplot}
\alias{score_barplot}
\title{barplot of processed score}
\usage{
score_barplot(top_markers, column = ".dot", f_list, n = 30)
}
\arguments{
\item{top_markers}{output of \code{\link[=top_markers]{top_markers()}}}

\item{column}{character, specify which column used as group label}

\item{f_list}{a named list of markers}

\item{n}{numeric, number of returned top genes for each group}
}
\value{
ggplot object
}
\description{
barplot of processed score
}
\examples{
data <- matrix(rnorm(100), 10, dimnames = list(1:10))
top_n <- top_markers(data, label = rep(c("A", "B"), 5))
score_barplot(top_n)
}
