% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartid-package.R
\docType{data}
\name{sim_sce_test}
\alias{sim_sce_test}
\title{scRNA-seq test data of 4 groups simulated by \code{splatter}.}
\format{
A SingleCellExperiment object of 100genes * 400 cells.
}
\source{
\code{\link[splatter:splatSimulate]{splatter::splatSimulate()}}
}
\usage{
data(sim_sce_test)
}
\value{
SingleCellExperiment
}
\description{
A SingleCellExperiment object containing 4 groups with each group up-regulated
DEGs saved in metadata.
}
\keyword{datasets}
