% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartid-package.R
\docType{package}
\name{smartid_Package}
\alias{smartid_Package}
\alias{smartid}
\title{Scoring and Marker Selection method based on modified TF-IDF}
\value{
Marker list and scores
}
\description{
\code{smartid} This package enables automated selection of group specific signature,
especially for rare population. The package is developed for generating
specific lists of signature genes based on TF-IDF modified methods. It
can also be used as a new gene-set scoring method or data transformation
method. Multiple visualization functions are implemented in this package.
}
\author{
Jinjin Chen \email{chen.j@wehi.edu.au}
}
\keyword{internal}
