% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_idf_iae_wrappers.R
\name{iae}
\alias{iae}
\title{standard inverse average expression}
\usage{
iae(expr, features = NULL, thres = 0)
}
\arguments{
\item{expr}{a matrix, features in row and cells in column}

\item{features}{vector, feature names or indexes to compute}

\item{thres}{numeric, cell only counts when expr > threshold, default 0}
}
\value{
a vector of iae score for each feature
}
\description{
standard inverse average expression
}
\examples{
data <- matrix(rpois(100, 2), 10, dimnames = list(1:10))
iae(data)
}
