% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_idf_iae_wrappers.R
\name{idf_rf}
\alias{idf_rf}
\title{labeled inverse cell frequency: relative frequency}
\usage{
idf_rf(expr, features = NULL, label, multi = TRUE, thres = 0)
}
\arguments{
\item{expr}{a matrix, features in row and cells in column}

\item{features}{vector, feature names or indexes to compute}

\item{label}{vector, group label of each cell}

\item{multi}{logical, if to compute based on binary (FALSE) or multi-class (TRUE)}

\item{thres}{numeric, cell only counts when expr > threshold, default 0}
}
\value{
a matrix of IDF score
}
\description{
labeled inverse cell frequency: relative frequency
}
\examples{
data <- matrix(rpois(100, 2), 10, dimnames = list(1:10))
smartid:::idf_rf(data, label = sample(c("A", "B"), 10, replace = TRUE))
}
