% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_markers.R
\name{top_markers_glm}
\alias{top_markers_glm}
\title{calculate group mean score using glm and order genes based on scores difference}
\usage{
top_markers_glm(
  data,
  label,
  n = 10,
  family = gaussian(),
  scale = TRUE,
  softmax = TRUE
)
}
\arguments{
\item{data}{matrix, features in row and samples in column}

\item{label}{vector, group labels}

\item{n}{numeric, number of returned top genes for each group}

\item{family}{family for glm, details in \code{\link[stats:glm]{stats::glm()}}}

\item{scale}{logical, if to scale data by row}

\item{softmax}{logical, if to apply softmax transformation on output}
}
\value{
a tibble with feature names, group labels and ordered processed scores
}
\description{
calculate group mean score using glm and order genes based on scores difference
}
\examples{
data <- matrix(rgamma(100, 2), 10, dimnames = list(1:10))
top_markers_glm(data, label = rep(c("A", "B"), 5))
}
